LayoutTests/imported/w3c/web-platform-tests/css/css-scroll-snap/scroll-initial-target/scroll-initial-target-aligns-with-snap-align.tentative-expected.txt


FAIL scroll-initial-target aligns with scroll-snap-align assert_equals: scroller is vertically scrolled to target expected 400 but got 0
